Refactor Jenkins credentials to be dynamic
Now an entire credentials object for the whole world is passed to the
API and invoked to get credentials for a request. This would allow
requesting arbitrary URLs sent back by the API, even if they pertain to
different Jenkins servers and require different credentials. Whether
Jenkins actually can do this is unknown.